Add video partial support
- Split promotional features methods into their own helper
- Add styles for videos partials links
- Videos partial
- Add the video title as a heading
- Use brand color for more videos link
- Use YouTube video URL as the YouTube link text
  To prevent duplication with the heading above it in no-js
  situations.
- Tests for helper
